Assign dispatchers to a per-network context obj
The only thing that the Network class manages by itself now is a
NetworkEventDispatcher and creation of the context class.
As a result, all events are now registered per-network and not globally,
except for the INIT_NETWORK_CTX event.

Decorators (for event handlers) are now wrapped in a convenience class.