Remove concept of a Channel class ...
... and store ChannelContexts directly in NetworkContext.channels dict.

The Channel class didn't hold any merit compared to its context class,
since most methods were to be added there anyway.

This also means that we create ChannelContexts when we join a channel
(instead of when we receive a message for a channel, which should never
happen before joining anyway).

It also means that non-channel PRIVMSG and NOTICE messages aren't
dispatched anymore, but their handling was really weird anyway.