Add type annotations for *all* functions
This is controlled with the `disallow_untyped_defs` option for mypy,
which is not set in the config by default.

Additionally, replace calls to `asyncio.ensure_future` with
`AbstractEventLoop.create_task` to ensure the return type is a Task.

Also use typing.NamedTuple in a couple places.

Comments have been added where workarounds were needed due to issues
with mypy or typeshed.