Initial rewrite of plugin and event logic
Plugins are now subclasses of some plugin class and get instantiated for each
network.
Events are marked with an `@event` decorator, which sets an attribute with a
HandlerInfo instance on them holding meta info like priority and event name.
All kinds of events were reduced to a single `Event` type, which accepts
a variable number of keyword arguments to be forwarded to handlers. As a
result, all events share the same namespace, but this shouldn't result in
conflicts.
Also, synchronous functions are now accepted, since most of the handlers we
had in our code base were running synchronously anyway.

Additional functionality (notably methods) are provided by new Plugin base
classes, which are defined in `shanghai.plugin_base` and are to be subclassed.
They are intended to function in a composition kind of way, so you can subclass
multiple of them when you need functionality provided by different ones.
Currently, the default plugin you want is most likely `MessagePlugin`.

Event handlers can request to append new events to the event queue with their
return value. They can also request coroutines to be scheduled in the event
loop and terminated (TODO) when the network disconnects. That way it is ensured
that the tasks will have finished once we cancel the process without having
each plugin manage its task manually.

All event handlers in network.py have been outsourced to core plugins, since
that file wouldn't be crawled for plugins.

Plugins currently not translated to the new systems (and not critical for
runtime):
- ctcp
- channel
- test

Tests were not adjusted and mypy wasn't run either.